Fixed project structure dependency between Core, Infrastructure and Web and Namespace Issues. Removed Circular Dependencies and fixed the docker build progress. Also created DI helper to centralize dependency registration.
