+++ EnronEmailSearch/Web/Dockerfile
@@ -5,18 +5,28 @@ EXPOSE 443
 
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
 WORKDIR /src
-COPY ["EnronEmailSearch/Indexer/Indexer.csproj", "EnronEmailSearch/Indexer/"]
+
+# Copy solution and project files first to restore dependencies
+COPY ["EnronEmailSearch.sln", "./"]
 COPY ["EnronEmailSearch/Core/Core.csproj", "EnronEmailSearch/Core/"]
 COPY ["EnronEmailSearch/Infrastructure/Infrastructure.csproj", "EnronEmailSearch/Infrastructure/"]
-RUN dotnet restore "EnronEmailSearch/Web/Web.csproj"
+# Change this line to match the actual location of your Web project
+COPY ["EnronEmailSearch/EnronEmailSearch.csproj", "EnronEmailSearch/"]
+
+# Restore dependencies before copying the rest of the source code
+RUN dotnet restore "EnronEmailSearch/EnronEmailSearch.csproj"
+
+# Copy the rest of the source code
 COPY . .
-WORKDIR "/src/EnronEmailSearch/Web"
-RUN dotnet build "Web.csproj" -c Release -o /app/build
+
+# Build the project
+WORKDIR "/src/EnronEmailSearch"
+RUN dotnet build "EnronEmailSearch.csproj" -c Release -o /app/build
 
 FROM build AS publish
-RUN dotnet publish "Web.csproj" -c Release -o /app/publish
+RUN dotnet publish "EnronEmailSearch.csproj" -c Release -o /app/publish
 
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "Web.dll"]
+ENTRYPOINT ["dotnet", "EnronEmailSearch.dll"]
