+++ docker-compose.microservice.yml
@@ -1,6 +1,4 @@
-﻿version: '3.8'
-
-services:
+﻿services:
   # Cleaner Service (Y-axis)
   cleaner-service:
     build:
